+++ src/components/spartan/index.ts
@@ -37,3 +37,4 @@ export * from './STab';
 export * from './STable';
 export * from './SDataTable';
 export * from './SInputIncrement';
+export * from './SInputIncrementBlock';