+++ src/components/spartan/SDataTable/types.ts
@@ -1,17 +1,36 @@
-import type { ColumnSort } from "@tanstack/table-core";
+import type { ColumnSort } from '@tanstack/table-core';
 
 export type TDataTableProps = {
-    cols: Record<string, string>[] | string[];
+    cols:
+        | {
+              id: string;
+              sortable?: boolean;
+              header?: string;
+              sortDescFirst?: boolean;
+          }[]
+        | string[];
     displayHeaderText?: (header: string) => string;
     data: unknown[];
     filtrable?: boolean;
     loading?: boolean;
-    pagination?: { page?: number; count: number; size?: number, sizes?: number[] };
-    sorting?: { availableColumns: string[]; currentSort?: ColumnSort, descFirst?: boolean };
+    pagination?: { page?: number; count: number; size?: number; sizes?: number[] };
+    sorting?: { currentSort?: ColumnSort; sortDescFirst?: boolean };
     containerClass?: string;
 };
 
 export type TDataTableEmits = {
     (event: 'paginationChange', value: Pick<Exclude<TDataTableProps['pagination'], undefined>, 'page' | 'size'>): void;
     (event: 'sortingChange', value: ColumnSort): void;
+    (
+        event: 'change',
+        value:
+            | {
+                  type: 'pagination';
+                  value: Pick<Exclude<TDataTableProps['pagination'], undefined>, 'page' | 'size'>;
+              }
+            | {
+                  type: 'sorting';
+                  value: ColumnSort;
+              },
+    ): void;
 };